Refactor: ui 단위 컴포넌트 분리, mock data 분리
